@@ -51,7 +51,7 @@ class TBodyCount extends Component {
   goTo(e) {
     e.preventDefault();
     this.props.showLoading();
-    const host = URL.parse(this.props.url.shortUrl).hostname;
+    const host = URL.parse(this.props.url.shortLink).hostname;
     Router.push(`/stats?id=${this.props.url.id}${`&domain=${host}`}`);
   }
 
@@ -77,14 +77,14 @@ class TBodyCount extends Component {
           )}
           <TBodyButton
             data-id={url.id}
-            data-host={URL.parse(url.shortUrl).hostname}
+            data-host={URL.parse(url.shortLink).hostname}
             onClick={showModal}
           >
             <Icon src="/images/trash.svg" />
           </TBodyButton>
         </Actions>
         <Modal show={this.state.showQrCodeModal} close={this.toggleQrCodeModal}>
-          <QRCode value={url.shortUrl} size={196} />
+          <QRCode value={url.shortLink} size={196} />
         </Modal>
       </Wrapper>
     );
@@ -98,7 +98,7 @@ TBodyCount.propTypes = {
     count: PropTypes.number,
     id: PropTypes.string,
     password: PropTypes.bool,
-    shortUrl: PropTypes.string,
+    shortLink: PropTypes.string,
   }).isRequired,
 };
 
