@@ -1,10 +1,11 @@
-exports.verifyMailText = `Thanks for creating an account on Kutt.it.
+/* eslint-disable max-len */
+export const verifyMailText = `Thanks for creating an account on Kutt.it.
 
 Please verify your email address using the link below.
 
 https://{{domain}}/{{verification}}`;
 
-exports.resetMailText = `A password reset has been requested for your account.
+export const resetMailText = `A password reset has been requested for your account.
 
 Please click on the button below to reset your password. There's no need to take any action if you didn't request this.
 
